Move JavaScript files to use webpacker
