+++ app/javascript/packs/application.js
@@ -1,1 +1,8 @@
 //=require bootstrap
+import {
+  createNewCargoItemForm,
+  editCargoItemForm
+} from "../views/fixture_cargo_add";
+
+createNewCargoItemForm();
+editCargoItemForm();
